axis_stream_to_pkt_backpressured: drain FIFOs when disabled
When the axis_stream_to_pkt_backpressured is enabled and get disabled,
some samples can remain in the sample FIFO. These samples will be
produced as the beginning of the output when the module is enabled
again, which is wrong, because it is old data.

To avoid this, a new state in the output state machine is introduced.
This state consumes data from the FIFOs without producing any output.
The state machine goes into this state when it finishes producing an
output packet and the module is disabled. It goes back to the header
state as soon as the module is enabled, so no samples can be lost
when re-enabling, because the input state machine takes more than 1
cycle to input data into the FIFO.

For this approach to work, the module needs to be disabled long
enough to drain the FIFOs completely (they are drained at a rate
of one word per cycle). This is usually the case if the enable is
controlled by software.