axis_stream_to_pkt_backpressured: allow disabling at any time
The input state machines only looked at enable in some states. This
potentially caused problems when the module was disabled, since the
state machines could get stuck in an unexpected state when the module
was re-enabled.

This treats !enable in the same way as rst in the input state machines,
simplifying the code and guaranteeing that the state machines are in
the initial state when the module is re-enabled.