axis_stream_to_pkt_backpressured: reset the FIFOs instead of draining
When the module gets disabled, to empty the FIFOs we reset them instead
of holding their TREADYs high to drain them. This empties the FIFOs
immediately rather than after several cycles.

For better timing closure, there is a delay of 1 cycle to clear the
FIFOs, since the comparison of the output state is registered. But this
is delay does not cause any problems.