axis_stream_to_pkt_backpressured: add missing reset to FIFO
In #28, some logic to reset the internal FIFOs when the
axis_stream_to_pkt_backpressured is disabled was added. However,
this 'fifos_reset' signal was not connected to sample_fifo_minimal
(it was only connected to time_fifo and sample_fifo). This caused
a samples beat to linger inside sample_fifo_minimal when the
module was disabled, which caused the first output packet after the
module was re-enabled to be one beat longer than it should (and
its length field in the DRaT header did not match the packet length
according to its TLAST).